+++ PyISY/isy.py
@@ -2,7 +2,6 @@
 import logging
 from threading import Thread
 
-from .climate import Climate
 from .clock import Clock
 from .configuration import Configuration
 from .connection import Connection
@@ -33,11 +32,8 @@ class ISY:
                           auto-reconnect to the event stream if the connection
                           is lost.
     :ivar auto_update: Boolean value that controls the class's subscription to
-                       the event stream that allows node, program, and climate
+                       the event stream that allows node, program
                        values to be updated automatically.
-    :ivar climate: Climate manager that holds all climate properties from the
-                   controller if the climate module is installed on the
-                   controller.
     :ivar connected: Read only boolean value indicating if the class is
                      connected to the controller.
     :ivar log: Logger used by the class and its children.
@@ -87,10 +83,6 @@ class ISY:
                 var_xml=self.conn.get_variables(),
             )
 
-            if self.configuration.get("Weather Information"):
-                self.climate = Climate(self, xml=self.conn.get_climate())
-            else:
-                self.climate = None
             # if self.configuration['Networking Module']:
             #     self.networking = NetworkResources(self,
             #       xml=self.conn.get_network())